Fix mismatch issues between int, size_t and uint32_t when compiling with warnings as errors

+++ code/client/api.c
@@ -924,7 +924,7 @@ HQAPI size_t GetItemModStruct(uint32_t item_id, uint32_t* buffer, size_t length)
     if (!item)
         goto leave;
     if (!buffer) {
-        written = item->mod_struct.size;
+        written = (int)item->mod_struct.size;
         goto leave;
     }
     if (length < item->mod_struct.size)
